Compact property pair list when too many deleted slots.
When properties are deleted some property pairs may hold only a single
property. The unused slot of the property pair is never reused unless
it is the first property pair. This patch introduces a function which
moves properties to the end of the property pair list and frees unused
property pairs in the beginning. The property order is not changed to
maintain creation order.

JerryScript-DCO-1.0-Signed-off-by: Zoltan Herczeg zherczeg.u-szeged@partner.samsung.com