Add new extension jerry-context
This is a user context manager that expands the number of pointers that
can be attached to a user context from one to a number specified at
compiled time.

It reuses the core interface whereby an init/deinit pair is passed and
only a getter is provided for the value created by the init and stored
in the context.

JerryScript-DCO-1.0-Signed-off-by: Gabriel Schulhof gabriel.schulhof@intel.com