Add new extension: context
This is a user context manager that expands the number of pointers that
can be attached to a user context from one to a number specified at
compile time.

It reuses the core interface whereby an init/deinit pair is passed and
only a getter is provided for the value created by the init and stored
in the context.

The context extension uses the section "jerryx_ctx_slots" in the binary
to store an array of structures defining each slot. Each structure has a
function pointer for the slot initializer and one for the deinitializer.

Macros are provided to make it easy to declare slots and to retrieve
the pointer at a given slot index from the current context.

JerryScript-DCO-1.0-Signed-off-by: Gabriel Schulhof gabriel.schulhof@intel.com