New extension: module
This extension provides two facilities: one for registering modules, and
one for registering module resolvers.

A module is defined as a global static structure containing a pointer
to a string which is the name of the module, and a pointer to a function
which will be called when an instance of the module is needed.

A module resolver is a function that accepts a string holding the name
of the module and returns a `jerry_value_t` if the module was found, or
zero if it was not. The resolver is expected to cache the
`jerry_value_t` it returns for a given string such that, if called with
the same string again, it will return the same `jerry_value_t`.

This extension provides a built-in module resolver which attempts to
load and cache modules that follow the above module definition.

Additionally, this extension provides an API for invoking all module
resolvers in sequence to attempt to resolve the name of a single module.
After one resolver returns a non-zero `jerry_value_t`, the API stops
iterating over the remaining resolvers and returns the `jerry_value_t`.
The API will return a `jerry_value_t` containing an error indicating
that the module was not found if it reaches the end of the list of
resolvers. The error it returns has an extra property `"moduleName"` the
value of which is a string containing the name of the module that the
API was asked to resolve.

JerryScript-DCO-1.0-Signed-off-by: Gabriel Schulhof gabriel.schulhof@intel.com