Add build option to generate jerryscript-snapshot-compiler.js
This adds a way to generate a JavaScript library / module
"jerryscript-snapshot-compiler.js" that can take a JS source string and produce
a JerryScript snapshot. This library is produced using the Emscripten compiler,
transpiling the C code that produces snapshots into JS. The advantage of such
a library as opposed to the binary "jerry" program is that as-is (without having
to recompile anything), it can be run on other platforms (i.e. cloud, mobile,
browser, etc.).

Use `tools/build.py --emscripten-snapshot-compiler ON` to generate the output.
Note that you need to have the Emscripten SDK activated.
Hint: run `source $PATH_TO_EMSDK/emsdk_env.sh` first.

JerryScript-DCO-1.0-Signed-off-by: Martijn The martijn.the@intel.com

Signed-off-by: Martijn The <martijn.the@intel.com>